change int8_t t int

+++ interface/JMEJetAnalyzer.h
@@ -62,8 +62,8 @@ class JMEJetAnalyzer : public JME::PhysicsObjectAnalyzer
         std::vector<int>& refpdgid = tree["refpdgid"].write<std::vector<int>>();
         std::vector<float>& refdrjt = tree["refdrjt"].write<std::vector<float>>();
         std::vector<float>& refarea = tree["refarea"].write<std::vector<float>>();
-        std::vector<int8_t>&   partonFlavor = tree["partonFlavor"].write<std::vector<int8_t>>();
-        std::vector<int8_t>&   hadronFlavor = tree["hadronFlavor"].write<std::vector<int8_t>>();
+        std::vector<int>&   partonFlavor = tree["partonFlavor"].write<std::vector<int>>();
+        std::vector<int>&   hadronFlavor = tree["hadronFlavor"].write<std::vector<int>>();
         std::vector<float>& jtarea = tree["jtarea"].write<std::vector<float>>();
         std::vector<float>& jec_toraw = tree["jec_toraw"].write<std::vector<float>>();
         std::vector<std::map<std::string, float>>& jec_factors = tree["jec_factors"].write<std::vector<std::map<std::string, float>>>();
